refactor: moved from teacher to core
